Pass siginfo_t->si_addr to the reginfo_init() function
On s390x, we need the si_addr from the siginfo_t to get to
the address of the illegal instruction (the PSW address in
the ucontext_t is already pointing to the next instruction
there). So let's prepare for that situation and pass the
si_addr to the reginfo_init() function everywhere.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-id: 20230914113311.379537-2-thuth@redhat.com
Signed-off-by: Peter Maydell <peter.maydell@linaro.org>