configure: Fix separate build tree detection
test -s is for "FILE exists and has a size greater than zero"
test -L is for "FILE exists and is a symbolic link (same as -h)"

If Makefile exists and is not a link, then we're not using a separate build tree.

In a separate build tree, Makefile hasn't been created yet, or it is a symbolic link. In either case we create or recreate the symbolic link and set BUILD_INC to include the separate build tree.