ppc64.risu: Fix lswi and stswi constraints
Problem test that should not happen:

    addi  r26, r1, 0
    stswi r0, r26, 0x19 ; rs=r0, ra=r26, rb=0x19

stswi stores ceiling(rb/4) registers, not just rs.
It should not store r1 because it is the stack pointer and therefore can differ between master and apprentice.
ra should not be stored because it is based on r1.
r13 should also not be stored for reasons?

lswi is similar but we allow loading ra.