ppc64.risu: Fix lswx and stswx constraints
First, choose a random number for xer[25-31]. Previously this was always 0.

stswx stores ceiling(xer[25-31]/4) registers, not just rs.
It should not store r1 because it is the stack pointer and therefore can differ between master and apprentice.
ra should not be stored because it is based on r1.
r13 should also not be stored for reasons?

lswx is similar but we allow loading ra.