Add risu opcodes setup begin and setup end
These opcodes will mark the beginning and end of the image's setup instructions. During execution of this range, a risu compare will not happen because that would probably fail since the setup has not completed yet. Also, illegal instructions are ignored during this time. This allows you to include vector instructions in the setup and still be able to use the test image on a CPU that doesn't have vector instructions.