ppc64: Fix reginfo_is_eq for illegal instructions
If an illegal instruction uses a register to access memory and that register is going to be restored after the instruction, then don't compare that register.
For this we need to store the next instruction. Make sure we don't read instructions outside the test image.
nip doesn't need to be 64 bit since it's an offset in the image and no-one is going to create a 4 GB image.