risugen: Add an option to set random number seed
Running risugen multiple times with the same input produces the same test image output.
If you create a test image with 10000 instructions and another with 20000 instructions, then the first 10000 instructions of the 2nd test image will be the same as in the first test image.
To fix that, we can change the random number seed.
Use a different value every time you create a test image to create a different test image.