ppc64: Propagate crf1 or fpscr# mask
fpscr and crf1 are affected by floating point calculations.

For fcmpo and fcmpu, any ccr field can be affected so that field should have the same mask as crf1.

For mcrfs, any ccr field can be affected so that field should have the same mask as the source field from fpscr.