risu: Allow run test image in emulation
Allow the risu test image to be executed in a different architecture/environment from the rest of the risu executable.
For example, risu could run the test image in an emulator and gather reginfo from the emulator state instead of from a signal handler's ucontext_t.

The included example is for the DingusPPC emulator which emulates a PowerPC 32-bit Big-Endian CPU. The emulator doesn't use a signal handler, so it will define NO_SIGNAL in config.h or as a compiler flag. DPPC is defined to indicate that the DingusPPC emulator will be used to execute the test image and gather the reginfo.

The reginfo struct is defined without using architecture specific types (such as the types used in a ucontext_t).

void *memblock is a pointer that the risu host uses to access memory in the guest's RAM. arch_ptr_t arch_memblock is the value that the guest architecture would use. arch_ptr_t is used for all guest pointers.

A risu master running inside a big-endian architecture produces trace files with big-endian values. A risu apprentice running in a little-endian host (with the test image running in an emulated big-endian environment) will endian swap as needed.
reginfo_host_to_arch and header_host_to_arch are used by the master to convert host (risu) to arch (test image) endian-ness.
header_arch_to_host and reginfo_arch_to_host are used by the apprentice to convert arch endian-ness to host endian-ness.
memory_blocks are copied and compared by byte, so no endian swapping is required.