risu: Don't update registers for illegal ops
Because !memory blocks may depend on some registers not changing (base addresses are based on the location of the master's memory block which may differ from the location of the apprentice's memory block). Illegal ops shouldn't change registers anyway.
Registers will get updated after the compare op. By that time any base register will be set to 0 or a value that is not based on a memory register.