ppc64: Compare link register
Init the link register to the PC in the test image.
Use PC relative comparison since apprentice and master will use different locations for the test image.
If the instruction being tested is mflr and the offset from the PC of the affected GPR in the master and the apprentice is the same, then ignore the difference and update the GPR apprentice value to match the master value.