risugen: Add ability to modify the instruction
Instead of randomly generating instructions until it matches a constraint, allow modifying the instruction so it matches a constraint. For example, if an instruction has a bit field that is 24 bits, but you want the bit field to contain only 1 of 16 values, it is quicker to transform the random 24 bit value into a 4 bit value than it is to continue generating values until it is one of the desired 16 values.